+++ src/tabular_trees/__init__.py
@@ -3,8 +3,11 @@
 import contextlib
 from importlib import metadata
 
-from .explain.prediction_decomposition import decompose_prediction
-from .explain.shapley_values import calculate_shapley_values
+from .explain.prediction_decomposition import (
+    PredictionDecomposition,
+    decompose_prediction,
+)
+from .explain.shapley_values import ShapleyValues, calculate_shapley_values
 from .lightgbm.booster_string import BoosterString
 from .lightgbm.editable_booster import EditableBooster
 from .lightgbm.lightgbm_tabular_trees import LightGBMTabularTrees
@@ -12,7 +15,7 @@ from .sklearn.sklearn_hist_tabular_trees import ScikitLearnHistTabularTrees
 from .sklearn.sklearn_tabular_trees import ScikitLearnTabularTrees
 from .trees import TabularTrees, export_tree_data
 from .validate import MonotonicConstraintResults, validate_monotonic_constraints
-from .xgboost.dump_parser import XGBoostParser
+from .xgboost.dump_parser import ParsedXGBoostTabularTrees, XGBoostParser
 from .xgboost.xgboost_tabular_trees import XGBoostTabularTrees
 
 # single source for version number is in the pyproject.toml
@@ -34,7 +37,9 @@ with contextlib.suppress(ImportError):
 
 __all__ = [
     "decompose_prediction",
+    "PredictionDecomposition",
     "calculate_shapley_values",
+    "ShapleyValues",
     "EditableBooster",
     "LightGBMTabularTrees",
     "BoosterString",
@@ -42,6 +47,7 @@ __all__ = [
     "ScikitLearnTabularTrees",
     "XGBoostTabularTrees",
     "XGBoostParser",
+    "ParsedXGBoostTabularTrees",
     "TabularTrees",
     "export_tree_data",
     "validate_monotonic_constraints",