+++ src/tabular_trees/sklearn/sklearn_tabular_trees.py
@@ -71,6 +71,11 @@ class ScikitLearnTabularTrees(BaseModelTabularTrees):
             GradientBoostingRegressor or Classifier tree data extracted from
             the .estimators_ attribute.
 
+        Returns
+        -------
+        trees : ScikitLearnTabularTrees
+            Model trees in tabular format.
+
         Examples
         --------
         >>> from sklearn.datasets import load_diabetes